Rename `KtUniversalisClient` to `KtUniversalis`
