+++ src/jsMain/kotlin/KtUniversalis.kt
@@ -18,14 +18,17 @@ import cloud.drakon.ktuniversalis.exception.ItemIdsException
 import cloud.drakon.ktuniversalis.exception.UniversalisException
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
-import io.ktor.client.engine.java.Java
+import io.ktor.client.engine.js.Js
 import io.ktor.client.plugins.DefaultRequest
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.client.request.get
 import io.ktor.serialization.kotlinx.json.json
+import kotlin.js.Promise
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.promise
 
-actual object KtUniversalisClient {
-    private val ktorClient = HttpClient(Java) {
+@JsExport actual object KtUniversalis {
+    private val ktorClient = HttpClient(Js) {
         install(ContentNegotiation) {
             json()
         }
@@ -39,11 +42,11 @@ actual object KtUniversalisClient {
      * Returns all data centers supported by the Universalis API
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getAvailableDataCenters(): Array<DataCenter> {
+    fun getAvailableDataCenters(): Promise<Array<DataCenter>> = GlobalScope.promise {
         val response = ktorClient.get("data-centers")
 
         if (response.status.value == 200) {
-            return response.body()
+            return@promise response.body()
         } else {
             throw UniversalisException()
         }
@@ -53,11 +56,11 @@ actual object KtUniversalisClient {
      * Returns the IDs and names of all worlds supported by the Universalis API
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getAvailableWorlds(): Array<World> {
+    fun getAvailableWorlds(): Promise<Array<World>> = GlobalScope.promise {
         val response = ktorClient.get("worlds")
 
         if (response.status.value == 200) {
-            return response.body()
+            return@promise response.body()
         } else {
             throw UniversalisException()
         }
@@ -72,11 +75,11 @@ actual object KtUniversalisClient {
      * @throws InvalidEntriesException `entries` must be between `0` and `200`
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getLeastRecentlyUpdatedItems(
+    fun getLeastRecentlyUpdatedItems(
         world: String? = null,
         dcName: String? = null,
         entries: Int? = null,
-    ): RecentlyUpdatedItems {
+    ): Promise<RecentlyUpdatedItems> = GlobalScope.promise {
         if (world == null && dcName == null) {
             throw InvalidWorldDcException()
         }
@@ -105,7 +108,7 @@ actual object KtUniversalisClient {
             }
 
         when (leastRecentlyUpdatedItems.status.value) {
-            200  -> return leastRecentlyUpdatedItems.body()
+            200  -> return@promise leastRecentlyUpdatedItems.body()
             404  -> throw InvalidWorldDcException()
             else -> throw UniversalisException()
         }
@@ -114,7 +117,7 @@ actual object KtUniversalisClient {
     /**
      * Returns the data currently shown on the market board for the requested array of item IDs and world or data center
      * @param worldDcRegion The world, data center, or region to retrieve data for. This may be an ID or a name. Regions should be specified as Japan, Europe, North-America, Oceania, China, or 中国.
-     * @param itemIds Array of item IDs to retrieve data for
+     * @param itemIds Array of item IDs to retrieve data for. Up to 100 item IDs can be provided.
      * @param listings The number of listings to return. By default, all listings will be returned.
      * @param entries The number of recent history entries to return. By default, a maximum of `5` entries will be returned.
      * @param noGst If the result should not have Gil sales tax (GST) factored in. GST is applied to all consumer purchases in-game, and is separate from the retainer city tax that impacts what sellers receive. By default, GST is factored in.
@@ -127,7 +130,7 @@ actual object KtUniversalisClient {
      * @throws InvalidWorldDcItemException The world/DC or item requested is invalid
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getMarketBoardCurrentData(
+    fun getMarketBoardCurrentData(
         worldDcRegion: String,
         itemIds: IntArray,
         listings: Int? = null,
@@ -137,7 +140,7 @@ actual object KtUniversalisClient {
         statsWithin: Int? = null,
         entriesWithin: Int? = null,
         fields: Array<String>? = null,
-    ): CurrentlyShown {
+    ): Promise<CurrentlyShown> = GlobalScope.promise {
         if (itemIds.size in 1 .. 100) {
             val marketBoardCurrentData =
                 ktorClient.get("$worldDcRegion/" + itemIds.joinToString(",")) {
@@ -167,7 +170,7 @@ actual object KtUniversalisClient {
                 }
 
             when (marketBoardCurrentData.status.value) {
-                200  -> return marketBoardCurrentData.body()
+                200  -> return@promise marketBoardCurrentData.body()
                 400  -> throw InvalidParameterException()
                 404  -> throw InvalidWorldDcItemException()
                 else -> throw UniversalisException()
@@ -188,13 +191,13 @@ actual object KtUniversalisClient {
      * @throws InvalidWorldDcItemException The world/DC or item requested is invalid
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getMarketBoardSaleHistory(
+    fun getMarketBoardSaleHistory(
         worldDcRegion: String,
         itemIds: IntArray,
         entriesToReturn: Int? = null,
         statsWithin: Int? = null,
         entriesWithin: Int? = null,
-    ): History {
+    ): Promise<History> = GlobalScope.promise {
         if (itemIds.size in 1 .. 100) {
             val marketBoardSaleHistory =
                 ktorClient.get("history/$worldDcRegion/" + itemIds.joinToString(",")) {
@@ -214,7 +217,7 @@ actual object KtUniversalisClient {
                 }
 
             when (marketBoardSaleHistory.status.value) {
-                200  -> return marketBoardSaleHistory.body()
+                200  -> return@promise marketBoardSaleHistory.body()
                 404  -> throw InvalidWorldDcItemException()
                 else -> throw UniversalisException()
             }
@@ -229,7 +232,7 @@ actual object KtUniversalisClient {
      * @throws InvalidWorldException The world requested is invalid
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getMarketTaxRates(world: String): TaxRates {
+    fun getMarketTaxRates(world: String): Promise<TaxRates> = GlobalScope.promise {
         val marketBoardTaxRates = ktorClient.get("tax-rates") {
             url {
                 parameters.append("world", world)
@@ -237,7 +240,7 @@ actual object KtUniversalisClient {
         }
 
         when (marketBoardTaxRates.status.value) {
-            200  -> return marketBoardTaxRates.body()
+            200  -> return@promise marketBoardTaxRates.body()
             404  -> throw InvalidWorldException()
             else -> throw UniversalisException()
         }
@@ -247,11 +250,11 @@ actual object KtUniversalisClient {
      * Returns an array of marketable item IDs
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getMarketableItems(): IntArray {
+    fun getMarketableItems(): Promise<IntArray> = GlobalScope.promise {
         val marketableItems = ktorClient.get("marketable")
 
         when (marketableItems.status.value) {
-            200  -> return marketableItems.body()
+            200  -> return@promise marketableItems.body()
             else -> throw UniversalisException()
         }
     }
@@ -265,11 +268,11 @@ actual object KtUniversalisClient {
      * @throws InvalidEntriesException `entries` must be between `0` and `200`
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getMostRecentlyUpdatedItems(
+    fun getMostRecentlyUpdatedItems(
         world: String? = null,
         dcName: String? = null,
         entries: Int? = null,
-    ): RecentlyUpdatedItems {
+    ): Promise<RecentlyUpdatedItems> = GlobalScope.promise {
         if (world == null && dcName == null) {
             throw InvalidWorldDcException()
         }
@@ -285,11 +288,11 @@ actual object KtUniversalisClient {
                     }
                     if (entries != null) {
                         when {
-                            entries <= 0 || entries > 200 -> throw InvalidEntriesException(
+                            (entries <= 0) || (entries > 200) -> throw InvalidEntriesException(
                                 "`entries` must be between `0` and `200`"
                             )
 
-                            else                          -> parameters.append(
+                            else                              -> parameters.append(
                                 "entries", entries.toString()
                             )
                         }
@@ -298,7 +301,7 @@ actual object KtUniversalisClient {
             }
 
         when (mostRecentlyUpdatedItems.status.value) {
-            200  -> return mostRecentlyUpdatedItems.body()
+            200  -> return@promise mostRecentlyUpdatedItems.body()
             404  -> throw InvalidWorldDcException()
             else -> throw UniversalisException()
         }
@@ -308,38 +311,41 @@ actual object KtUniversalisClient {
      * Returns the total upload counts for each client application that uploads data to Universalis
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getUploadCountsByUploadApplication(): Array<SourceUploadCount> {
-        val uploadCountsByUploadApplication =
-            ktorClient.get("extra/stats/uploader-upload-counts")
+    fun getUploadCountsByUploadApplication(): Promise<Array<SourceUploadCount>> =
+        GlobalScope.promise {
+            val uploadCountsByUploadApplication =
+                ktorClient.get("extra/stats/uploader-upload-counts")
 
-        when (uploadCountsByUploadApplication.status.value) {
-            200  -> return uploadCountsByUploadApplication.body()
-            else -> throw UniversalisException()
+            when (uploadCountsByUploadApplication.status.value) {
+                200  -> return@promise uploadCountsByUploadApplication.body()
+                else -> throw UniversalisException()
+            }
         }
-    }
 
     /**
      * Returns the world upload counts and proportions of the total uploads for each world
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getUploadCountsByWorld(): Map<String, WorldUploadCount> {
-        val getUploadCountsByWorld = ktorClient.get("extra/stats/world-upload-counts")
+    fun getUploadCountsByWorld(): Promise<Map<String, WorldUploadCount>> =
+        GlobalScope.promise {
+            val getUploadCountsByWorld =
+                ktorClient.get("extra/stats/world-upload-counts")
 
-        when (getUploadCountsByWorld.status.value) {
-            200  -> return getUploadCountsByWorld.body()
-            else -> throw UniversalisException()
+            when (getUploadCountsByWorld.status.value) {
+                200  -> return@promise getUploadCountsByWorld.body()
+                else -> throw UniversalisException()
+            }
         }
-    }
 
     /**
      * Returns the number of uploads per day over the past 30 days
      * @throws UniversalisException The Universalis API returned an unexpected return code
      */
-    suspend fun getUploadsPerDay(): UploadCountHistory {
+    fun getUploadsPerDay(): Promise<UploadCountHistory> = GlobalScope.promise {
         val getUploadsPerDay = ktorClient.get("extra/stats/upload-history")
 
         when (getUploadsPerDay.status.value) {
-            200  -> return getUploadsPerDay.body()
+            200  -> return@promise getUploadsPerDay.body()
             else -> throw UniversalisException()
         }
     }
